Added generic AudioSampleValue and AudioData types. These classes will be used for all audio data.
Implemented AudioSampleValue and AudioData for the au file format thereby
extending the number of different encodings AuFile can cope with (and making
it easily extendable).

Made the SampleValue interface a bit thinner (by implementing Radius and
isNeighbour() in SampleValue, not in derived classes).

+++ steghide/src/BmpFile.cc
@@ -112,9 +112,9 @@ void BmpFile::replaceSample (const SamplePos pos, const SampleValue* s)
 	}
 }
 
-unsigned int BmpFile::getSamplesPerEBit() const
+unsigned short BmpFile::getSamplesPerEBit() const
 {
-	unsigned int retval = 0 ;
+	unsigned short retval = 0 ;
 	switch (getBitCount()) {
 		case 1: case 4: case 8: {
 			retval = 2 ;
