+++ steghide/src/JpegFile.h
@@ -44,7 +44,7 @@ class JpegFile : public CvrStgFile {
 	unsigned long getNumSamples (void) const ;
 	void replaceSample (const SamplePos pos, const SampleValue* s) ;
 	SampleValue* getSampleValue (SamplePos pos) const ;
-	unsigned int getSamplesPerEBit (void) const ;
+	unsigned short getSamplesPerEBit (void) const ;
 
 	private:
 	/// the frame of the jpeg file
