+++ steghide/src/WavFile.h
@@ -42,7 +42,7 @@ class WavFile : public CvrStgFile {
 	unsigned long getNumSamples (void) const ;
 	void replaceSample (const SamplePos pos, const SampleValue* s) ;
 	SampleValue* getSampleValue (SamplePos pos) const ;
-	unsigned int getSamplesPerEBit (void) const ;
+	unsigned short getSamplesPerEBit (void) const ;
 
 	unsigned short getBitsPerSample (void) const ;
